backend/requirements.txt
django
django-webpack-loader
channels_redis
channels
djangorestframework